refactor(use context selection): move hook into its own folder
