+++ src/context-selection/category-option-combo-selector-bar-item/use-category-combination.js
@@ -1,5 +1,5 @@
 import { useQuery } from 'react-query'
-import { useDataSetId } from '../use-context-selection.js'
+import { useDataSetId } from '../use-context-selection/index.js'
 
 const QUERY_CATEGORY_COMBINATION = {
     dataSet: {
