+++ src/context-selection/category-option-combo-selector-bar-item/use-on-dependent-params-change.js
@@ -3,7 +3,7 @@ import {
     useDataSetId,
     useOrgUnitId,
     usePeriodId,
-} from '../use-context-selection.js'
+} from '../use-context-selection/index.js'
 
 export default function useOnDependentParamsChange(cb) {
     const initialExecution = useRef(true)
