+++ src/context-selection/context-selection/context-selection.js
@@ -5,7 +5,7 @@ import { DataSetSelectorBarItem } from '../data-set-selector-bar-item/index.js'
 import { OrgUnitSetSelectorBarItem } from '../org-unit-selector-bar-item/index.js'
 import { PeriodSelectorBarItem } from '../period-selector-bar-item/index.js'
 import { SectionFilterSelectorBarItem } from '../section-filter-selector-bar-item/index.js'
-import { useClearEntireSelection } from '../use-context-selection.js'
+import { useClearEntireSelection } from '../use-context-selection/index.js'
 import useShouldHideClearButton from './use-should-hide-clear-button.js'
 
 export default function ContextSelector() {
