+++ src/context-selection/index.js
@@ -1,2 +1,2 @@
-export { useContextSelection } from './use-context-selection.js'
+export { useContextSelection } from './use-context-selection/index.js'
 export { ContextSelection } from './context-selection/index.js'
