+++ src/context-selection/org-unit-selector-bar-item/disabled-tooltip.js
@@ -2,7 +2,7 @@ import i18n from '@dhis2/d2-i18n'
 import { Tooltip } from '@dhis2/ui'
 import PropTypes from 'prop-types'
 import React from 'react'
-import { useDataSetId } from '../use-context-selection.js'
+import { useDataSetId } from '../use-context-selection/index.js'
 
 export default function DisabledTooltip({ children }) {
     const [dataSetId] = useDataSetId()
