+++ src/context-selection/org-unit-selector-bar-item/use-data-set-org-unit-paths.js
@@ -1,5 +1,5 @@
 import { useQuery } from 'react-query'
-import { useDataSetId } from '../use-context-selection.js'
+import { useDataSetId } from '../use-context-selection/index.js'
 
 const QUERY_DATA_SET_ORG_UNIT_PATHS = {
     dataSet: {
