+++ src/context-selection/period-selector-bar-item/period-selector-bar-item.js
@@ -2,7 +2,7 @@ import i18n from '@dhis2/d2-i18n'
 import { SelectorBarItem } from '@dhis2/ui'
 import React, { useEffect, useState } from 'react'
 import { usePeriod } from '../../shared/index.js'
-import { useDataSetId, usePeriodId } from '../use-context-selection.js'
+import { useDataSetId, usePeriodId } from '../use-context-selection/index.js'
 import computeMaxYear from './compute-max-year.js'
 import PeriodMenu from './period-menu.js'
 import useDataSetPeriodType from './use-data-set-period-type.js'
