+++ src/context-selection/section-filter-selector-bar-item/section-filter-selector-bar-item.js
@@ -2,7 +2,7 @@ import i18n from '@dhis2/d2-i18n'
 import { SelectorBarItem } from '@dhis2/ui'
 import React, { useState } from 'react'
 import { MenuSelect } from '../menu-select/index.js'
-import { useSectionFilter } from '../use-context-selection.js'
+import { useSectionFilter } from '../use-context-selection/index.js'
 import useDataSetSectionsInfo from './use-data-set-sections-info.js'
 import useOnDependentParamsChange from './use-on-dependent-params-change.js'
 import useSelectorBarItemValue from './use-selector-bar-item-value.js'
