+++ src/context-selection/section-filter-selector-bar-item/use-on-dependent-params-change.js
@@ -4,7 +4,7 @@ import {
     useDataSetId,
     useOrgUnitId,
     usePeriodId,
-} from '../use-context-selection.js'
+} from '../use-context-selection/index.js'
 
 export default function useOnDependentParamsChange(cb) {
     const initialExecution = useRef(true)
