+++ src/context-selection/section-filter-selector-bar-item/use-selector-bar-item-value.js
@@ -1,5 +1,5 @@
 import i18n from '@dhis2/d2-i18n'
-import { useSectionFilter } from '../use-context-selection.js'
+import { useSectionFilter } from '../use-context-selection/index.js'
 import useDataSetSectionsInfo from './use-data-set-sections-info.js'
 
 export default function useSelectorBarItemValue() {
